Add twitter integration.
Allows to set a twitter handle for the current chat.
Checks that twitter timeline for new tweets every 2 minutes (if a handle is set)
Say it in chat. Also allows to ask for the last tweet with "+latest"

This is a work in progress. I don't like that it has weechat-specific code in the
BotTwitterMixin. But currently I haven't found a pretty way to hook this up.

Forgot the import. Oops